v0/pyproject.toml
[project]
name = "prod"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "zwang038@gold.ac.uk", email = "zwang038@gold.ac.uk" }
]
dependencies = [
    "django==4.2.*",
    "django-stubs==4.2.7",
    "irc>=20.3.1",
    "pydantic>=2.6.4",
    "djangorestframework>=3.15.1",
    "gevent>=24.2.1",
    "django-cors-headers>=4.3.1",
    "django-extensions>=3.2.3",
    "environs[django]>=11.0.0",
    "whitenoise>=6.6.0",
    "django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot.git@develop",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
]
readme = "README.md"
requires-python = "== 3.10.13"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "isort>=5.13.2",
    "black>=24.3.0",
    "ipython>=8.22.2",
    "pytest-django>=4.8.0",
    "factory-boy>=3.3.0",
    "pytest-sugar>=1.0.0",
]
